% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie_characters.R
\docType{data}
\name{movie_characters}
\alias{movie_characters}
\title{Star Wars Characters}
\format{
An object of class \code{character} of length 14.
}
\usage{
movie_characters
}
\description{
Named vector of star wars characters
}
\details{
\code{movie_characters} is a named character vector of characters,
with each element a character string of variable length specifying
an ASCII star wars character. Note that some have unicode characters that won't
play well on some operating systems.
}
\examples{
movie_characters['baby_yoda']
movie_characters['R2D2']
movie_characters[['R2D2']]
}
\keyword{datasets}
