% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Messages with star wars characters}
\usage{
say(
  what = "May the force be with you",
  by = "stormtrooper",
  what_color = NULL,
  by_color = NULL
)
}
\arguments{
\item{what}{(character) What do you want to say?}

\item{by}{(character) Select a some character, baby_yoda, R2D2, black_one, at_at_walker,
darth_vader, resistance, at_st_walker, yoda, ackbar, darth_maul, stormtrooper and stormtrooper_body,
starwars, starwars_2, millenium_falcon.
Alternatively, use "random" to have your message spoken by a random character.}

\item{what_color}{(character or crayon function) One or more
\href{https://github.com/r-lib/crayon#256-colors}{\code{crayon}}-suported
text color(s) or
\href{https://github.com/r-lib/crayon#styles}{\code{crayon style function}}
to color \code{what}. You might try \code{colors()} or \code{?rgb} for ideas.
Use "rainbow" for c("red", "orange", "yellow", "green", "blue", "purple").}

\item{by_color}{(character or crayon function) One or more
\href{https://github.com/r-lib/crayon#256-colors}{\code{crayon}}-suported
text color(s) or
\href{https://github.com/r-lib/crayon#styles}{\code{crayon style function}}
to color \code{who}. Use "rainbow" for
\code{c("red", "orange", "yellow", "green", "blue", "purple")}}
}
\value{
A message with your selected character
}
\description{
Messages with star wars characters
}
\examples{
say()
say("I'm your father", by = "darth_vader", by_color = "black")

}
